+++ packages/eyes-sdk-core/lib/new/abort.js
@@ -1,10 +1,6 @@
 function makeAbort({eyes}) {
   return async function abort() {
-    let results = await eyes.abort()
-
-    if (!results) return []
-
-    return results.map(result => (result ? result.toJSON() : null))
+    return await eyes.abort() || []
   }
 }